+++ src/main/java/org/delia/assoc/ManyToManyVisitor.java
@@ -1,0 +1,8 @@
+package org.delia.assoc;
+
+import org.delia.rule.rules.RelationRuleBase;
+import org.delia.type.DStructType;
+
+public interface ManyToManyVisitor {
+	void visit(DStructType structType, RelationRuleBase rr);
+}
